Change type of port variables from int to uint16

+++ internal/services/frontend/frontend.go
@@ -40,7 +40,7 @@ import (
 // data fields.
 //
 type Server struct {
-	port         int
+	port         uint16
 	rootFilePath string
 
 	handler     http.Handler
